+++ src/test/java/br/uff/ic/graphmatching/GraphMatchingTest.java
@@ -44,24 +44,24 @@ public class GraphMatchingTest {
     @Test
     public void testIsSimilar() {
         System.out.println("isSimilar");
-        equals(1.00, true);
-        almostEquals(0.75, true);
-        halfEquals(0.5, true);
-        quarterEquals(0.25, true);
-        notEquals(0.00, true);
+        equals(1.00f, true);
+        almostEquals(0.75f, true);
+        halfEquals(0.5f, true);
+        quarterEquals(0.25f, true);
+        notEquals(0.00f, true);
         
-        equals(0.75, true);
-        almostEquals(0.85, false);
-        halfEquals(0.75, false);
-        quarterEquals(0.5, false);
-        notEquals(0.10, false);
+        equals(0.75f, true);
+        almostEquals(0.85f, false);
+        halfEquals(0.75f, false);
+        quarterEquals(0.5f, false);
+        notEquals(0.10f, false);
         
-        differentTypes(1.00, false);
-        differentAttributes(0.5, true);
-        differentAttributes(0.51, false);
+        differentTypes(1.00f, false);
+        differentAttributes(0.5f, true);
+        differentAttributes(0.51f, false);
     }
     
-    public void Comparing(double threshold, boolean expResult, 
+    public void Comparing(float threshold, boolean expResult, 
             String a11, String a12, 
             String a21, String a22, 
             String e1, String e2, String e3, String e4) {
@@ -104,7 +104,7 @@ public class GraphMatchingTest {
         assertEquals(expResult, result);
     }
     
-    public void ComparingDefault(double threshold, boolean expResult, 
+    public void ComparingDefault(float threshold, boolean expResult, 
             String a11, String a12, 
             String a21, String a22, 
             String e3, String e4) {
@@ -141,7 +141,7 @@ public class GraphMatchingTest {
         assertEquals(expResult, result);
     }
     
-    public void equals(double threshold, boolean expResult) {
+    public void equals(float threshold, boolean expResult) {
         System.out.println("Equal");
         Comparing(threshold, expResult, "asd", "bvF", "2012-05-24T10:00:02", "2012-05-24T10:00:01", "asd, bvf", "2000", "50%", "300%");
         Comparing(threshold, expResult, "8", "10", "2012-05-24T10:00:02", "2012-05-24T10:00:01", "2", "2000", "50%", "300%");
@@ -149,29 +149,29 @@ public class GraphMatchingTest {
         ComparingDefault(threshold, expResult, "8", "10", "-10", "-10", "50%", "300%");
     }
     
-    public void almostEquals(double threshold, boolean expResult) {
+    public void almostEquals(float threshold, boolean expResult) {
         System.out.println("AlmostEqual");
         Comparing(threshold, expResult, "asd", "AsD", "2012-05-24T10:00:01", "2012-05-24T10:00:02", "0", "2000", "50%", "0%");
         ComparingDefault(threshold, expResult, "8", "10", "7", "10", "50%", "300%");
     }
     
-    public void halfEquals(double threshold, boolean expResult) {
+    public void halfEquals(float threshold, boolean expResult) {
         System.out.println("HalfEqual");
         Comparing(threshold, expResult, "asd", "AsD", "2012-05-24T10:00:01", "2012-05-24T10:00:02", "0", "2000", "0%", "0%");
         ComparingDefault(threshold, expResult, "8", "10", "7", "10", "50%", "0");
     }
     
-    public void quarterEquals(double threshold, boolean expResult) {
+    public void quarterEquals(float threshold, boolean expResult) {
         System.out.println("notEqual");
         Comparing(threshold, expResult, "asd", "AsD", "2012-05-24T10:00:01", "2012-05-24T10:00:02", "0", "0", "0%", "0%");
     }
     
-    public void notEquals(double threshold, boolean expResult) {
+    public void notEquals(float threshold, boolean expResult) {
         System.out.println("notEqual");
         Comparing(threshold, expResult, "asd", "AsDD", "2012-05-24T10:00:01", "2012-05-24T10:00:02", "0", "0", "0%", "0%");
     }
     
-    public void differentTypes(double threshold, boolean expResult) {
+    public void differentTypes(float threshold, boolean expResult) {
         System.out.println("Different Types");
         Vertex v1 = new ActivityVertex("v1", "test", "0");
         Vertex v2 = new EntityVertex("v2", "test", "0");
@@ -188,7 +188,7 @@ public class GraphMatchingTest {
         assertEquals(expResult, result);
     }
     
-    public void differentAttributes(double threshold, boolean expResult) {
+    public void differentAttributes(float threshold, boolean expResult) {
         System.out.println("Different Types");
         Vertex v1 = new ActivityVertex("v1", "test", "0");
         Vertex v2 = new ActivityVertex("v2", "test", "0");
@@ -255,7 +255,7 @@ public class GraphMatchingTest {
         av2 = new GraphAttribute("a6", "6");
         v2.addAttribute(av2);
         
-        GraphMatching instance = new GraphMatching(null, 0.0);
+        GraphMatching instance = new GraphMatching(null, 0.0f);
         
         Vertex expResult = new ActivityVertex(v1.getID() + "_" + v2.getID() + " (Merged)", v1.getLabel() + "_" + v2.getLabel(), v1.getTimeString());
         GraphAttribute aResult;
@@ -325,7 +325,7 @@ public class GraphMatchingTest {
     public void testUpdateEdges() {
         System.out.println("updateEdges");
         Collection<Edge> edges = new ArrayList<Edge>();
-        GraphMatching instance = new GraphMatching(null, 0.0);
+        GraphMatching instance = new GraphMatching(null, 0.0f);
         
         Vertex v1 = new ActivityVertex("v1", "test01", "0");
         Vertex v2 = new ActivityVertex("v2", "test02", "0");
